lib/python-extension/pyproject.toml
[tool.poetry]
authors = ["Brandon Bjelland <brandonjacob@google.com>"]
description = ""
name = "python-clouddeploy-extension"
packages = [{include = "clouddeploy_extension"}]
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
fastapi = "^0.92.0"
google-cloud-deploy = "^1.6.1"
google-cloud-logging = "^3.5.0"
google-cloud-secret-manager = "^2.15.1"
pydantic = "^1.10.4"
python = "^3.10"
python-dotenv = "^0.21.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
black = "^23.1.0"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.pytest.ini_options]
addopts = "-ra -s"
# log_cli = 'True'
# log_cli_level = 'INFO'
minversion = "7.0"
testpaths = [
  "tests",
]
